backends/: should use size_t n_regs_mapping for *_sample_sp_pc
These should have been size_t all along, matching the public
libdwfl_stacktrace dwflst_sample_getframes interface. Fix.

* backends/aarch64_initreg_sample.c (aarch64_sample_sp_pc): Fixup
  argument type for n_regs_mapping.
* backends/i386_initreg_sample.c (i386_sample_sp_pc): Fixup argument
  type for n_regs_mapping.
* backends/libebl_PERF_FLAGS.h (generic_sample_sp_pc): Fixup argument
  type for n_regs_mapping, adjust loop index var accordingly.
* backends/x86_64_initreg_sample.c (x86_64_sample_sp_pc): Fixup
  argument type for n_regs_mapping.
* libebl/ebl-hooks.h (sample_sp_pc): Fixup argument type for
  n_regs_mapping.

+++ backends/aarch64_initreg_sample.c
@@ -39,7 +39,7 @@
 
 bool
 aarch64_sample_sp_pc (const Dwarf_Word *regs, uint32_t n_regs,
-		      const int *regs_mapping, uint32_t n_regs_mapping,
+		      const int *regs_mapping, size_t n_regs_mapping,
 		      Dwarf_Word *sp, Dwarf_Word *pc)
 {
   return generic_sample_sp_pc (regs, n_regs, regs_mapping, n_regs_mapping,
